Fixed Seating API hookups

+++ panel/app/ManageUsers.jsx
@@ -20,7 +20,13 @@ class ManageUsers extends React.Component {
 
   //TRANSFER DATA BETWEEN COMPONENTS
   //passed in as a callback prop to DynamicForms component
-  transfer(data) {this.setState({ data })};
+  transfer(data) {
+    if (!Array.isArray(data)) {//check if the data came back as an array
+      this.setState({data: [data]});//if not, store it as an array so it can be processed accordingly
+    } else {//otherwise, it's already an array
+      this.setState({ data });//store it in the state
+    }
+  };
 
   //FORM CONTROLS
   methodMenu(e) {this.setState({ method: e.target.value })}
@@ -31,7 +37,7 @@ class ManageUsers extends React.Component {
       .then(resp => resp.json())
       .then(data => this.setState({cityData: data}))
       .catch(err => console.error(err));
-  }//
+  }
 
   render() {
     return (
@@ -96,7 +102,7 @@ class DynamicForms extends React.Component {
     };
     this.props.callback([]);//Reset the "Data" state in ManageUsers before updating it
     //GET request for the input data
-    fetch('https://savi-travel.com:'+config.port+'/api/users'+searchTerm())
+    fetch('https://savi-travel.com:'+config.port+'/api/users'+searchTerm(), {mode: 'no-cors'})
       .then(resp => resp.json())
       .then(data => this.props.callback(data))//sends the data up to the ManageUsers component
       .catch(err => console.error(err));
@@ -222,9 +228,9 @@ class UserData extends React.Component {
       mdn: this.props.data.mdn,
       country: this.props.data.country,
       city: this.props.data.city,
-      cityId: 1,
       type: this.props.data.type,
-      seats: 1
+      origType: this.props.data.type,
+      seats: 0
     };
 
     //METHOD BINDINGS
@@ -236,7 +242,7 @@ class UserData extends React.Component {
     this.countryForm = this.countryForm.bind(this);
     this.cityForm = this.cityForm.bind(this);
     this.typeForm = this.typeForm.bind(this);
-
+    this.seatsForm = this.seatsForm.bind(this);
   }
 
   //FORM CONTROLS
@@ -244,43 +250,48 @@ class UserData extends React.Component {
   emailForm(e) {this.setState({userEmail: e.target.value})};
   mdnForm(e) {this.setState({mdn: e.target.value})};
   countryForm(e) {this.setState({country: e.target.value})};
-  cityForm(e) {this.setState({city: e.target.value})};
+  cityForm(e) {//specialized cityForm method to handle multiple data returns
+    //process the inbound data...
+    let data = e.target.value.split(',');
+    //set both states
+    this.setState({city: data[0], cityId: data[1]});
+  };
   typeForm(e) {this.setState({type: e.target.value})};
+  seatsForm(e) {this.setState({seats: e.target.value})};
   //toggle edit option for individual users
   toggleEdit() {this.setState({ edit: true })};
 
+  //Events that take place when the "Save" button is clicked
   saveHandler(e) {
-    if (this.props.data.type !== this.state.type) {//if the original props "type" doesn't match the state "type"...
-      if (this.state.type === "Tourist") {//if the new type is "Tourist"...
-        //delete the employee entry for this userId
-        fetch('https://savi-travel.com:'+config.port+'/api/employees', {
-          method: 'DELETE',
-          headers: {
-            'Accept': 'application/json',
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({
-            userId: this.props.data.id
-          })
-        });
-      } else if (this.state.type === "Driver" || this.state.type === "Tour Guide") {//otherwise, the new state must be either "Tour Guide" or "Driver"
-        //create a new employee entry for this userId
-        console.log("CREATING NEW EMPLOYEE")
-        fetch('https://savi-travel.com:'+config.port+'/api/employees', {
-          method: 'POST',
-          headers: {
-            'Accept': 'application/json',
-            'Content-Type': 'application/json'
-          },
-          body: JSON.stringify({
-            type: this.state.type,
-            rating: 3, //need to change this to be more dynamic
-            seats: this.state.seats,
-            userId: this.props.data.id,
-            cityId: 1//need to accurately pair the cityId with the cityName
-          })
-        });
-      }
+    if (this.state.type === "Tourist" && this.state.origType !== "Tourist") {//if the new type is "Tourist"...
+      //delete the employee entry for this userId
+      fetch('https://savi-travel.com:'+config.port+'/api/employees', {
+        method: 'DELETE',
+        headers: {
+          'Accept': 'application/json',
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({
+          userId: this.props.data.id
+        })
+      });
+    } else if ((this.state.type === "Driver" || this.state.type === "Tour Guide") && this.state.origType === "Tourist") {//otherwise, the new state must be either "Tour Guide" or "Driver"
+      //create a new employee entry for this userId
+      fetch('https://savi-travel.com:'+config.port+'/api/employees', {
+        method: 'POST',
+        headers: {
+          'Accept': 'application/json',
+          'Content-Type': 'application/json'
+        },
+        body: JSON.stringify({
+          type: this.state.type,
+          rating: 3,
+          seats: this.state.seats,
+          userId: this.props.data.id,
+          city: this.state.city
+        })
+      });
+    } else {//in any other case...
       //send a request to update the employee
       fetch('https://savi-travel.com:'+config.port+'/api/employees/'+this.props.data.id, {
         method: 'PUT',
@@ -290,12 +301,12 @@ class UserData extends React.Component {
         },
         body: JSON.stringify({
           type: this.state.type,
-          rating: 3,//need to change this to be more dynamic,
+          rating: 3,
           seats: this.state.seats,
-          cityId: 1//need to accurately pair the cityId with the cityName
+          city: this.state.city
         })
       });
-    }//... and then do the following in every case
+    }
 
     //PUT REQUEST FOR UPDATING USER INFORMATION
     fetch('https://savi-travel.com:'+config.port+'/api/users/'+this.state.userAuthId, {
@@ -314,7 +325,18 @@ class UserData extends React.Component {
         city: this.state.city
       })
     });
-    this.setState({ edit: false });//Toggle state back to false when save button is clicked
+    this.setState({ edit: false, origType: this.state.type });//Toggle state back to false when save button is clicked
+  }
+
+  componentWillMount() {
+    if (this.props.data.type !== "Tourist") {//If the user is not a tourist...
+      //get their employee data and set their "seat" value to match what is returned
+      fetch('https://savi-travel.com:'+config.port+'/api/employees?userId='+this.props.data.id, {
+        method: 'GET'
+      }).then(resp => resp.json())
+      .then(data => this.setState({seats: data.seats}))
+      .catch(err => console.error(err));
+    }
   }
 
   render() {
@@ -356,6 +378,22 @@ class UserData extends React.Component {
                     <option value="Tour Guide">Tour Guide</option>
                   </select>
                 </div>
+                {(()=>{
+                  if (this.state.type === "Driver") {
+                    return (
+                      <div>
+                      Available Seats:
+                        <select onChange={this.seatsForm} value={this.state.seats}>
+                          {[0,1,2,3,4,5,6,7,8,9,10].map((item, i) => {
+                            return (
+                              <option key={i} value={item}>{item}</option>
+                            )
+                          })}
+                        </select>
+                      </div>
+                    )
+                  }
+                })()}
               </div>
             )
           } else {
@@ -367,6 +405,15 @@ class UserData extends React.Component {
                 <div>Country: {this.state.country}</div>
                 <div>City: {this.state.city}</div>
                 <div>Status: {this.state.type}</div>
+                {(()=>{
+                  if (this.state.type === "Driver") {
+                    return (
+                      <div>
+                        Available Seats: {this.state.seats}
+                      </div>
+                    )
+                  }
+                })()}
               </div>
             )
           }
